+++ apps/studio/src/routes/editor/EditPanel/ChatTab/ChatInput.tsx
@@ -12,6 +12,7 @@ import { AnimatePresence } from 'motion/react';
 import { useState } from 'react';
 import { DraftContextPill } from './ContextPills/DraftContextPill';
 import { DraftImagePill } from './ContextPills/DraftingImagePill';
+import { compressImage } from '@/lib/utils';
 
 export const ChatInput = observer(() => {
     const editorEngine = useEditorEngine();
@@ -123,22 +124,6 @@ export const ChatInput = observer(() => {
         };
         reader.readAsDataURL(file);
     };
-
-    async function compressImage(file: File): Promise<string | undefined> {
-        const options = {
-            maxSizeMB: 1,
-            maxWidthOrHeight: 1024,
-        };
-
-        try {
-            const compressedFile = await imageCompression(file, options);
-            const base64URL = imageCompression.getDataUrlFromFile(compressedFile);
-            console.log(`Image size reduced from ${file.size} to ${compressedFile.size} (bytes)`);
-            return base64URL;
-        } catch (error) {
-            console.error('Error compressing image:', error);
-        }
-    }
 
     const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
         e.preventDefault();
